refactor(pdf): isolate visual candidate matching
